Add RewardPanel updates
